Move lua_ecs files to Scene
