+++ Sprocket/Scene/Systems/basic_systems.cpp
@@ -1,8 +1,8 @@
 #include "basic_systems.h"
 
 #include <Sprocket/Scene/scene.h>
+#include <Sprocket/Scene/lua_ecs.h>
 #include <Sprocket/Scripting/lua_input.h>
-#include <Sprocket/Scripting/lua_ecs.h>
 #include <Sprocket/Scripting/lua_maths.h>
 #include <Sprocket/Scripting/LuaScript.h>
 #include <Sprocket/Utility/input_store.h>
